feat: refactor for better CI/CD using ansible
