+++ .github/workflows/old_build.yml
@@ -1,25 +1,17 @@
 name: build & deploy
 
 on:
-  push:
-    branches: ['main', 'ci']
   workflow_dispatch:
 
 jobs:
+  # build server and web
+  # dockerize each
+  # push to container registry
   build_push:
     name: Build and push image
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-
-      - name: Install Nix
-        uses: DeterminateSystems/nix-installer-action@v8
-
-      - name: Setup Nix cache
-        uses: DeterminateSystems/magic-nix-cache-action@v2
-
-      - name: Build Docker image
-        run: nix build -L --accept-flake-config .#container
 
       - name: Load Docker image
         run: docker load < result
@@ -27,22 +19,15 @@ jobs:
       - name: Tag Docker image
         run: docker tag ${{ vars.IMAGE_NAME }}:latest ${{ secrets.REGISTRY }}/${{ vars.IMAGE_NAME }}:latest
 
-      - run: sudo apt install wireguard
-      - run: echo "${{ secrets.WIREGUARD_PRIVATE_KEY }}" > privatekey
-      - run: sudo ip link add dev wg0 type wireguard
-      - run: sudo ip address add dev wg0 ${{ vars.WIREGUARD_CLIENT_IP }} peer ${{ vars.WIREGUARD_SERVER_IP }}
-      - run: sudo wg set wg0 listen-port ${{ vars.WIREGUARD_PORT }} private-key privatekey peer "${{ vars.WIREGUARD_SERVER_PUBLIC_KEY }}" allowed-ips ${{ vars.WIREGUARD_ALLOWED_IPS }} endpoint "${{ secrets.SERVER_IP }}":${{ vars.WIREGUARD_PORT }}
-      - run: sudo ip link set up dev wg0
-
-      - name: Add wg hosts to /etc/hosts
-        run: |
-            sudo echo "${{ vars.WIREGUARD_SERVER_IP }} ${{ secrets.SERVER_PRIVATE_URL }}" | sudo tee -a /etc/hosts
-
       - name: Push Docker image
         run: docker push ${{ secrets.REGISTRY }}/${{ vars.IMAGE_NAME }}:latest
 
       - run: rm privatekey
 
+  # connect to VPS
+  # pull new images from registry
+  # update caddy configuration
+  # restart docker-compose 
   deploy:
     name: Deploy
     needs: build_push
