+++ .github/workflows/push.yml
@@ -1,0 +1,40 @@
+name: Dockerize Go Application
+
+on:
+  workflow_run:
+    workflows: ["Build Go Application"]
+    types:
+      - completed
+
+jobs:
+  dockerize:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+
+    - name: Download build artifact
+      uses: actions/download-artifact@v3
+      with:
+        name: counter
+        path: counter
+
+    - name: Set up Docker Buildx
+      uses: docker/setup-buildx-action@v2
+
+    - name: Log in to Docker Hub
+      uses: docker/login-action@v2
+      with:
+        registry: ghcr.io
+        username: ${{ github.actor }}
+        password: ${{ secrets.GHCR_TOKEN_WRITE }}
+
+    - name: Build and push Docker image
+      uses: docker/build-push-action@v4
+      with:
+        context: .
+        file: ./docker/server/Dockerfile
+        push: true
+        tags: "v3lmx/osc_server:latest"
+
