--- Dockerfile
@@ -1,3 +1,0 @@
-FROM alpine:latest
-COPY build/counter /counter
-ENTRYPOINT ["/counter"]
