+++ ansible/playbook.yml
@@ -1,0 +1,9 @@
+- name: My first play
+  hosts: staging_01
+  tasks:
+   - name: Ping my hosts
+     ansible.builtin.ping:
+
+   - name: Print message
+     ansible.builtin.debug:
+       msg: Hello world
