+++ ansible/requirements.yml
@@ -1,0 +1,9 @@
+collections:
+- name: community.general
+- name: community.docker
+- name: devsec.hardening
+
+roles:
+  - name: githubixx.ansible_role_wireguard
+    src: https://github.com/githubixx/ansible-role-wireguard.git
+    version: 17.0.0
