+++ ansible/wireguard/setup.yml
@@ -1,0 +1,34 @@
+---
+- name: Set up WireGuard server
+  hosts: all
+  become: true
+  vars_files:
+    - vars.yml
+
+  tasks:
+    - name: Update apt cache
+      apt:
+        update_cache: yes
+
+    - name: Install WireGuard
+      apt:
+        name: wireguard
+        state: present
+
+    - name: Ensure WireGuard directory exists
+      file:
+        path: /etc/wireguard
+        state: directory
+        mode: '0700'
+
+    - name: Copy WireGuard configuration
+      template:
+        src: templates/wg0.conf.j2
+        dest: /etc/wireguard/wg0.conf
+        mode: '0600'
+
+    - name: Enable and start WireGuard service
+      systemd:
+        name: wg-quick@wg0
+        enabled: yes
+        state: started
