+++ compose.yml
@@ -1,0 +1,17 @@
+version: '3'
+services:
+  proxy:
+    image: "caddy:latest"
+    ports:
+      - "443:4443"
+      - "80:8080"
+
+  server:
+    image: "ghcr.io/v3lmx/onesinglecounter/server:latest"
+    ports:
+      - "10001:10001"
+
+  web:
+    image: "ghcr.io/v3lmx/onesinglecounter/web:latest"
+    ports:
+      - "10002:10002"
