+++ docker/server/Dockerfile
@@ -1,0 +1,3 @@
+FROM scratch
+COPY ./server/counter .
+CMD ["/counter"]
