--- mprocs.log
@@ -1,1 +1,0 @@
-ERROR [mprocs::error] Error: channel closed
