+++ server/internal/core/cond.go
