+++ svelte/.vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "f819542e",
+  "configHash": "f4737d6d",
+  "lockfileHash": "e3b0c442",
+  "browserHash": "5b1cb5c8",
+  "optimized": {},
+  "chunks": {}
+}
