+++ web-svelte/package.json
