+++ web-svelte/src/lib/count.svelte.js
