+++ web-svelte/src/main.js
