+++ web-svelte/src/vite-env.d.ts
