+++ web-svelte/svelte.config.js
