+++ web-svelte/vite.config.js
