+++ web-vanilla/script.js
